+++ stm8/uart.h
@@ -28,6 +28,6 @@ void uart_write_str(const char *str);
 void uart_write_int(uint16_t val);
 void uart_write_int32(uint32_t val);
 void uart_write_fixed_point(uint32_t val);
-void uart_write_millivalue(uint16_t val);
+void uart_write_millivalue(uint32_t val);
 void uart_drive(void);
 void uart_flush_writes(void);
